Rework some Deferred code to fix a couple of bugs.
Throwing an exception in promise code apparently does not cause the
promise to have its state set to "rejected", but leaves it "pending" and
so prevents client code from handling the error (since the exception
also can't be caught, since it is thrown by code that is executed
asynchronously).

Chaining promises with "done" doesn't seem to work properly, but
chaining with "then" does. When using "done" to chain B to A, if B gets
rejected then A can still end up "resolved". When using "then" things
work as expected. I guess this is more likely a problem with my
understanding of jQuery, rather than a problem with jQuery itself.

Also, having keystone.authenticateDeferred being a Deferred object
constructed with a beforeStart parameter which was a function that would
resolve/reject keystone.authenticateDeferred seems like a bug. Instead,
chain promises with "then".